Ignore pathologically large depths longer than strlen
Closes #66

For example, for `simdjson_decode('{}', true, 1000000000)`,
we only need a depth that's at least 2 to get identical results.

PHP programmers may assume that a large depth is a safe way to avoid
depth errors because it was safe for `json_decode`

To avoid allocating too much virtual memory and wasting page table
entries or potentially fatal out of memory errors,
reduce depth to a safe value behaving the same way if the requested depth is
larger than the allocated depth, larger than the string length, and
exceeds 100000.
Do this in a way that reduces reallocations.